deleteFile improve

+++ controllers/fileHandler/deleteFile.js
@@ -1,17 +1,26 @@
-import path from "path";
+import pathTo from "path";
 import fs from "fs";
+import { codeService } from "../../services/code.service.js";
 
 const deleteFile = async (req, res) => {
-  const filePath = path.join(`/DownVuer/Vue/src/Node/${req.body.title}`);
+  const decoded = req.body.link.split("/")[5];
+  const { path } = codeService()
+    .getStorage()
+    .find((x) => x.code === decoded);
 
-  fs.unlink(filePath, (err) => {
-    console.log("usuwam");
-    if (err) {
-      console.error(err);
-    } else {
-      res.status(200).send("Delete");
-    }
-  });
+  if (path) {
+    const filePath = pathTo.join(`${path}`);
+
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        console.error(err);
+      } else {
+        console.log("Deleted " + filePath);
+      }
+    });
+  } else {
+    res.status(404).send("Not found");
+  }
 };
 
 export { deleteFile };
